pem on jenkins thru cfn params

+++ deploy.nando.automation.demo.sh
@@ -1,13 +1,23 @@
 #!/bin/bash
 
+clear
 echo
 echo "Nando Automation Demo Launch Script"
 echo
 keyName="nando-demo"
+cfnFile="file://cloudformation.json"
+existingStack=$(aws cloudformation describe-stacks --stack-name $keyName 2> /dev/null)
+if [[ $existingStack == *CREATE_COMPLETE* ]]; then 
+	echo
+	echo "Stack \"$keyName\" exists. Please delete manually before executing this script."
+	echo
+	echo
+	exit 666
+fi
 existingKeypair=$(aws ec2 describe-key-pairs --key-name $keyName) 
 if [ ! -z "$existingKeypair" ]; then 
 	echo
-	echo "Deleting existing $keyName keypair:... $existingKeypair"; 
+	echo "Deleting existing $keyName keypair: $existingKeypair"; 
 	aws ec2 delete-key-pair --key-name $keyName
 	echo
 fi
@@ -15,9 +25,36 @@ echo
 echo "Creating $keyName keypair:"
 echo
 privateKeyValue=$(aws ec2 create-key-pair --key-name $keyName)
-echo $privateKeyValue
 echo
+echo "Launching stack:"
+echo
+aws cloudformation create-stack --stack-name $keyName --template-body $cfnFile --parameters "ParameterKey=PrivateKey,ParameterValue=$privateKeyValue"
 
-aws cloudformation create-stack --stack-name nando-demo --template-body file://cloudformation.json --parameters ParameterKey=privateKey,ParameterValue=$privateKeyValue
-#while (!$(aws cloudformation describe-stacks --stack-name nando-demo|grep CREATE_COMPLETE)); do sleep 10; done
-#echo "\n\nThe Demo has been deployed.\n\n"
+complete=0
+second=0
+while [ "$complete" -ne 1 ]; do
+	stackStatus=$(aws cloudformation describe-stacks --stack-name $keyName)
+	if [[ $stackStatus == *ROLLBACK* ]]; then
+		echo
+		echo "FAILURE"
+		echo
+		echo $stackStatus
+		echo
+		echo
+		exit 666
+	elif [[ $stackStatus != *CREATE_COMPLETE* ]]; then 
+		sleep 1; 
+		echo -n "."; 
+		let seconds=seconds+1
+	else 
+		echo
+		echo $stackStatus
+		echo
+		complete=1; 
+	fi
+done
+echo
+echo
+echo "The Demo has been deployed in $seconds seconds."
+echo
+echo
